+++ src/penn_chime/charts.py
@@ -505,7 +505,7 @@ def build_ppe_descriptions(
 
     if asterisk:
         messages.append(
-            "_* The max is at the upper bound of the data, and therefore may not be the actual max_")
+            "* _The max is at the upper bound of the data, and therefore may not be the actual max_")
     return "  \n".join(messages)
 
 def build_staffing_descriptions(
@@ -544,7 +544,7 @@ def build_staffing_descriptions(
 
     if asterisk:
         messages.append(
-            "_* The max is at the upper bound of the data, and therefore may not be the actual max_")
+            "* _The max is at the upper bound of the data, and therefore may not be the actual max_")
     return "  \n".join(messages)
 
 def build_table(
@@ -556,7 +556,11 @@ def build_table(
     table_df = df[np.mod(df.day, modulo) == 0].copy()
     table_df.date = table_df.date.dt.strftime(DATE_FORMAT)
     table_df.rename(labels)
-    if "masks_n95" in table_df.columns:
+    # Remove non_icu_after_icu column when printing
+    if "non_icu_after_icu" in table_df.columns:
+        table_df = table_df.drop(columns="non_icu_after_icu")
+
+    if "masks_n95_icu" in table_df.columns:
         table_df = table_df[[
             "day",
             "date",
@@ -579,7 +583,7 @@ def build_table(
             "other_ppe_non_icu",
             "other_ppe_icu",
         ]]
-    if "nurses_total" in table_df.columns:
+    if "nurses_icu" in table_df.columns:
         table_df = table_df[[
             "day",
             "date",