+++ changelog.md
@@ -5,9 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-## [2.0.0] 2020-05-04
+## [2.1.0] 2020-05-05
 ### Added 
 - Emprical forecast mode.  Leverages county infection and population data to forecast future infections.
+### Changed
+- Changed how ICU length of stay is represented.
 
 ## [1.7.1] - 2020-04-14
 ### Added